[Chore] rename to `app.ts`
